feat(exampl): comm examples
extends the interface.rs to include a minimal function to keep stronghold running
